+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ledgerhq/live-cli
 
+## 24.17.3
+
+### Patch Changes
+
+- Updated dependencies [[`486e860`](https://github.com/LedgerHQ/ledger-live/commit/486e860cbad0ca9ba54e231c9d42739e0ccee662)]:
+  - @ledgerhq/live-common@34.31.2
+
 ## 24.17.2
 
 ### Patch Changes