+++ libs/live-dmk-mobile/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ledgerhq/live-dmk-mobile
 
+## 0.4.1
+
+### Patch Changes
+
+- [#10055](https://github.com/LedgerHQ/ledger-live/pull/10055) [`486e860`](https://github.com/LedgerHQ/ledger-live/commit/486e860cbad0ca9ba54e231c9d42739e0ccee662) Thanks [@mbertin-ledger](https://github.com/mbertin-ledger)! - DMK fixes for LLM
+
+- Updated dependencies [[`486e860`](https://github.com/LedgerHQ/ledger-live/commit/486e860cbad0ca9ba54e231c9d42739e0ccee662)]:
+  - @ledgerhq/live-dmk-shared@0.8.1
+
 ## 0.4.0
 
 ### Minor Changes