+++ tests/dummy-wallet-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dummy-wallet-app
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [[`486e860`](https://github.com/LedgerHQ/ledger-live/commit/486e860cbad0ca9ba54e231c9d42739e0ccee662)]:
+  - @ledgerhq/live-common@34.31.2
+
 ## 0.5.1
 
 ### Patch Changes